알림 채널 앱 실행시 생성하도록 수정
