+++ app/src/main/java/com/foundy/hansungnotification/HansungNotificationApp.kt
@@ -1,8 +1,11 @@
 package com.foundy.hansungnotification
 
-import android.Manifest
 import android.app.Application
-import com.eazypermissions.coroutinespermission.PermissionManager
+import android.app.NotificationChannel
+import android.app.NotificationManager
+import android.content.Context
+import android.os.Build
+import com.foundy.presentation.R
 import com.google.firebase.database.ktx.database
 import com.google.firebase.ktx.Firebase
 import dagger.hilt.android.HiltAndroidApp
@@ -13,5 +16,16 @@ class HansungNotificationApp : Application() {
     override fun onCreate() {
         super.onCreate()
         Firebase.database.setPersistenceEnabled(true)
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            val notificationManager =
+                getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+            val channel = NotificationChannel(
+                getString(R.string.firebase_keyword_noti_channel_id),
+                getString(R.string.keyword_notification),
+                NotificationManager.IMPORTANCE_DEFAULT
+            )
+            notificationManager.createNotificationChannel(channel)
+        }
     }
 }
